Ai/deploy/requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart
pyyaml
pandas
